requirements.txt
Django==1.9.2
Pillow==3.2.0
argh==0.26.1
dj-database-url==0.4.0
dj-static==0.0.6
django-appconf==1.0.1
django-avatar==3.0.0
django-dotenv==1.4.1
django-extensions==1.6.1
django-filter==0.12.0
django-getenv==1.3.1
django-registration==2.0.4
django-registration-redux==1.4
six==1.10.0
static3==0.6.1
whitenoise==2.0.6
